Update tests for fields class
